fullstack/server/requirements.txt
browser-use==0.1.41
pyperclip==1.9.0
gradio==5.27.0
json-repair
langchain-mistralai==0.2.4
MainContentExtractor==0.0.4
langchain-ibm==0.3.10
langchain_mcp_adapters==0.0.9
langgraph==0.3.34
langchain-community

# FastAPI authentication system dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
pytest>=7.3.1
httpx>=0.24.0
cloudinary>=1.36.0
alembic==1.16.4